Chain recovery.

+++ tools/cluster/cluster.go
@@ -521,7 +521,7 @@ func (clu *Cluster) start() error {
 	return nil
 }
 
-func (clu *Cluster) KillNodeProcess(nodeIndex int) error {
+func (clu *Cluster) KillNodeProcess(nodeIndex int, gracefully bool) error {
 	if nodeIndex >= len(clu.waspCmds) {
 		return fmt.Errorf("[cluster] Wasp node with index %d not found", nodeIndex)
 	}
@@ -531,8 +531,17 @@ func (clu *Cluster) KillNodeProcess(nodeIndex int) error {
 		return nil
 	}
 
-	if err := wcmd.cmd.Process.Kill(); err != nil {
-		return err
+	if gracefully {
+		if err := wcmd.cmd.Process.Signal(os.Interrupt); err != nil {
+			return err
+		}
+		if _, err := wcmd.cmd.Process.Wait(); err != nil {
+			return err
+		}
+	} else {
+		if err := wcmd.cmd.Process.Kill(); err != nil {
+			return err
+		}
 	}
 
 	clu.waspCmds[nodeIndex] = nil
@@ -694,7 +703,7 @@ func (clu *Cluster) stopNode(nodeIndex int) {
 	}
 	fmt.Printf("[cluster] Sending shutdown to wasp node %d\n", nodeIndex)
 
-	err := clu.KillNodeProcess(nodeIndex)
+	err := clu.KillNodeProcess(nodeIndex, true)
 	if err != nil {
 		fmt.Println(err)
 	}
